moved parser into its own module
changing the parsing table into its own class. Parsing is done not on
PNodes but on PTable_entries which rely on position on table and not on
copies of given element. This way 1) overhead is reduced as there is no
copy 2) scrapping every tree out of the table is possible, as we need to
mark somehow which nodes have been visited