trying to add sppf into complete parser tests
this makes little sense, the specific sppf nodes which are created depend on the algorithm and perhaps it is not a good idea to hard-code them. We only care about the results that are returned as roots, everything other should remain untested.

work on extracting the sppfs that are currently generated and saving them as the expected results has been done. Leaving it as it is

REMOVED TCS WITH EMPTY SYMBOLS